refactor: unify SQLModel schemas/models (leverage SQLModel), restructure for maintainability and versioned API, clean main via master_router, improve DB structure
