+++ app/utils/jwt.py
@@ -6,7 +6,7 @@ from jwt.exceptions import InvalidTokenError
 from sqlmodel import SQLModel
 from app.database.database import SessionDep
 from app.utils.hashing import verify_password
-from app.models.user import User
+from app.api.models.user import User
 from datetime import datetime, timedelta, timezone
 from app.config.config import SECRET_KEY, ACCESS_TOKEN_EXPIRE_MINUTES, ALGORITHM
 
@@ -70,7 +70,7 @@ async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)],
 async def get_current_active_user(
     current_user: Annotated[User, Depends(get_current_user)],
 ):
-    if current_user.disabled:
+    if not current_user.is_active:
         raise HTTPException(status_code=400, detail="Inactive user")
     return current_user
